only remove or copy config if not already in place

+++ run.sh
@@ -5,8 +5,8 @@ set -e
 TARGETDIR="/fancontroller"
 
 echo "*** Copy config file..."
-cp "/tmp/fan_control.yaml" "$TARGETDIR/fan_control.yaml"
-rm "/tmp/fan_control.yaml"
+if [ ! -f "$TARGETDIR/fan_control.yaml" ]; then cp "/tmp/fan_control.yaml" "$TARGETDIR/fan_control.yaml" fi
+if [ -f "/tmp/fan_control.yaml" ]; then rm "/tmp/fan_control.yaml" fi
 
 echo "*** Start fan controller..."
 "/fancontroller/venv/bin/python3" -u "/fancontroller/fan_control.py" --config "/fancontroller/fan_control.yaml"
